third_party/WebKit/LayoutTests/external/wpt/XMLHttpRequest/send-content-type-string-expected.txt

This is a testharness.js-based test.
PASS XMLHttpRequest: send() - Content-Type
PASS XMLHttpRequest: send() - Content-Type 1
FAIL XMLHttpRequest: send() - Content-Type 2assert_equals: expected "text/html;charset=UTF-8" but got "application/xml;charset=UTF-8"
Harness: the test ran to completion.
